Make default file context match with named transitions
Some of files in the /etc/insights-client directory do not inherit
the default insights_client_etc_t, but are assigned insights_client_etc_rw_t
instead. The list in the default file context specification needs to
match with named file transitions for insights_client_t as well as in
the insights_client_filetrans_named_content() interface.

Resolves: rhbz#2107363